+++ HuffmanArchiver/HuffmanArchiver/Source.cpp
@@ -92,7 +92,7 @@ std::pair<std::vector<uint8_t>, std::string> read_unzip(std::string path) {
 	FileReader reader(path);
 	std::vector<std::pair<uint8_t, BitVector>> table = read_table(reader);
 
-	std::vector<std::pair<std::pair<size_t, int>, uint8_t>> nums;
+	std::vector<std::pair<std::pair<uint64_t, int>, uint8_t>> nums;
 	for (int i = 0; i < table.size(); ++i) {
 		nums.push_back(std::make_pair(std::make_pair(table[i].second.value(), table[i].second.size()), table[i].first));
 	}
@@ -103,7 +103,7 @@ std::pair<std::vector<uint8_t>, std::string> read_unzip(std::string path) {
 
 	std::string extension = "";
 	int zeros = 0;
-	size_t curr_bits = 0;
+	uint64_t curr_bits = 0;
 	int curr_sz = 0;
 	while (buffer.size() + extension.size() + 2 != sz) {
 		bool bit = reader.get_next_bit();
@@ -145,8 +145,8 @@ void run(std::string operation, std::string file_path) {
 
 		std::reverse(file_buff.begin(), file_buff.end());
 		file_buff.push_back(0);
-		size_t last_backslash = file_path.rfind('\\');
-		size_t ext_index = file_path.rfind('.');
+		uint64_t last_backslash = file_path.rfind('\\');
+		uint64_t ext_index = file_path.rfind('.');
 		if (ext_index > last_backslash) {
 			std::string ext = file_path.substr(ext_index + 1);
 			for (int i = ext.size() - 1; i >= 0; --i) {
